Removed guard on GetUserId to allow workshop to flow without authorization.

+++ save-points/03-show-order-status/BlazingPizza.Server/PizzaApiExtensions.cs
@@ -19,6 +19,10 @@ public static class PizzaApiExtensions
                 // We're storing at most one subscription per user, so delete old ones.
                 // Alternatively, you could let the user register multiple subscriptions from different browsers/devices.
                 var userId = GetUserId(context);
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 var oldSubscriptions = db.NotificationSubscriptions.Where(e => e.UserId == userId);
                 db.NotificationSubscriptions.RemoveRange(oldSubscriptions);
 
@@ -49,10 +53,9 @@ public static class PizzaApiExtensions
 
     }
 
-    public static string GetUserId(HttpContext context)
+    public static string? GetUserId(HttpContext context)
     {
         string? result = context.User.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (result is null) throw new UnauthorizedAccessException("User claim was not found.");
         return result;
     }
